requirements.txt
pymysql
pysocks
pyyaml
requests==2.12.5
sqlalchemy